fix: small refactor
