+++ tests/test_apps.py
@@ -1,7 +1,7 @@
 from django.utils import autoreload
 from django.utils.autoreload import autoreload_started
 
-from next.urls import get_pages_directories_for_watch
+from next.pages import get_pages_directories_for_watch
 from next.utils import NextStatReloader
