+++ tests/test_pages.py
@@ -3225,7 +3225,7 @@ class TestGetLayoutDjxPathsForWatch:
         (tmp_path / "a" / "layout.djx").write_text("<div>a</div>")
         (tmp_path / "a" / "b").mkdir()
         (tmp_path / "a" / "b" / "layout.djx").write_text("<div>b</div>")
-        with patch("next.urls.get_pages_directories_for_watch") as mock_watch:
+        with patch("next.pages.get_pages_directories_for_watch") as mock_watch:
             mock_watch.return_value = [tmp_path]
             result = get_layout_djx_paths_for_watch()
         assert len(result) == 2
@@ -3235,7 +3235,7 @@ class TestGetLayoutDjxPathsForWatch:
 
     def test_returns_empty_when_no_layout_djx(self, tmp_path) -> None:
         """Returns empty set when no layout.djx under pages dirs."""
-        with patch("next.urls.get_pages_directories_for_watch") as mock_watch:
+        with patch("next.pages.get_pages_directories_for_watch") as mock_watch:
             mock_watch.return_value = [tmp_path]
             result = get_layout_djx_paths_for_watch()
         assert result == set()
@@ -3243,7 +3243,7 @@ class TestGetLayoutDjxPathsForWatch:
     def test_swallows_oserror_on_rglob_layout(self, tmp_path) -> None:
         """When rglob raises OSError (e.g. permission), log and return partial result."""
         with (
-            patch("next.urls.get_pages_directories_for_watch") as mock_watch,
+            patch("next.pages.get_pages_directories_for_watch") as mock_watch,
             patch.object(Path, "rglob", side_effect=OSError(13, "Permission denied")),
         ):
             mock_watch.return_value = [tmp_path]
@@ -3260,7 +3260,7 @@ class TestGetTemplateDjxPathsForWatch:
         (tmp_path / "x" / "template.djx").write_text("x")
         (tmp_path / "x" / "y").mkdir()
         (tmp_path / "x" / "y" / "template.djx").write_text("y")
-        with patch("next.urls.get_pages_directories_for_watch") as mock_watch:
+        with patch("next.pages.get_pages_directories_for_watch") as mock_watch:
             mock_watch.return_value = [tmp_path]
             result = get_template_djx_paths_for_watch()
         assert len(result) == 2
@@ -3270,7 +3270,7 @@ class TestGetTemplateDjxPathsForWatch:
 
     def test_returns_empty_when_no_template_djx(self, tmp_path) -> None:
         """Returns empty set when no template.djx under pages dirs."""
-        with patch("next.urls.get_pages_directories_for_watch") as mock_watch:
+        with patch("next.pages.get_pages_directories_for_watch") as mock_watch:
             mock_watch.return_value = [tmp_path]
             result = get_template_djx_paths_for_watch()
         assert result == set()
@@ -3278,7 +3278,7 @@ class TestGetTemplateDjxPathsForWatch:
     def test_swallows_oserror_on_rglob_template(self, tmp_path) -> None:
         """When rglob raises OSError (e.g. permission), log and return partial result."""
         with (
-            patch("next.urls.get_pages_directories_for_watch") as mock_watch,
+            patch("next.pages.get_pages_directories_for_watch") as mock_watch,
             patch.object(Path, "rglob", side_effect=OSError(13, "Permission denied")),
         ):
             mock_watch.return_value = [tmp_path]
